+++ Makefile
@@ -1,6 +1,7 @@
 MAIN = srcs/main/prompt.c \
 		srcs/parsing/lexer.c \
-		srcs/parsing/list_helpers.c
+		srcs/parsing/list_helpers.c \
+		srcs/parsing/strings.c
 
 CC = cc
 CFLAGS = -Wall -Wextra -Werror
@@ -10,9 +11,12 @@ OBJS = ${MAIN:.c=.o}
 
 all: ${NAME}
 
+# Remove the flags at compilation for testing
+
 $(NAME):
+		@echo "Compiling"
 		@make all -C libft
-		$(CC) $(MAIN) -L./libft -lft -L/usr/local/lib -I/usr/local/include -lreadline -o ${NAME}
+		$(CC) $(MAIN) -L./libft -lft -L/usr/local/lib -I/usr/local/include -lreadline -o ${NAME} && echo "Everything compiled!"
 
 %.o: %.c
 	${CC} $(CFLAGS) -c $<
@@ -27,3 +31,5 @@ fclean: clean
 re: fclean all
 
 .PHONY: all clean fclean re
+
+.SILENT: all clean fclean re
