+++ includes/parsing.h
@@ -3,17 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   parsing.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jbarbay <jbarbay@student.42.fr>            +#+  +:+       +#+        */
+/*   By: jbarbay <jbarbay@student.42singapore.sg    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/15 22:28:38 by jbarbay           #+#    #+#             */
-/*   Updated: 2024/01/18 18:42:34 by jbarbay          ###   ########.fr       */
+/*   Updated: 2024/01/19 17:25:16 by jbarbay          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef PARSING_H
 #define PARSING_H
 
-#define STRING 0
+#define WORD 0
 #define SINGLE 1
 #define DOUBLE 2
 #define PIPE 3
@@ -35,10 +35,11 @@ t_token	*get_tokens(char *line);
 void	create_token(int type, char *str, t_token **list);
 int		process_quote(char **line, int quote, t_token **list);
 int		process_pipe(t_token **list);
-int		process_variable(char **line, t_token **list);
-int		process_input(char **line, char *c, t_token **list);
+int		process_env_var(char **line, t_token **list);
+int		process_redirect(char **line, char *c, t_token **list);
 int		error_tokens(int error, t_token **list);
-
+int		process_string(char **line, t_token **list);
+int		find_index(const char *s, int c);
 
 // Token list helpers
 void	print_tokens(t_token *token);
